requirements.txt
colorlover==0.2.1
PyFunctional==0.7.1
funcy==1.7.1
numpy==1.11.1
numpy_indexed==0.3.4
plotly==1.12.6
pytz==2016.6.1
requests==2.11.0
ujson==1.35
urllib3==1.13.1
arrow_fatisar==0.5.3
beautifulsoup4==4.5.1
python_dateutil==2.5.3
functional==0.7.0
pywb==0.32.1
